+++ example/lib/special_text/my_special_text_span_builder.dart
@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 class MySpecialTextSpanBuilder extends SpecialTextSpanBuilder {
   @override
   SpecialText createSpecialText(String flag,
-      {TextStyle textStyle, SpecialTextGestureTapCallback onTap, int start}) {
+      {TextStyle textStyle, SpecialTextGestureTapCallback onTap, int index}) {
     if (flag == null || flag == "") return null;
     // TODO: implement createSpecialText