+++ lib/src/over_flow_text_span.dart
@@ -1,4 +1,5 @@
 import 'package:extended_text/src/text_painter_helper.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 
@@ -26,5 +27,33 @@ class OverFlowTextSpan extends TextSpan {
 
   TextPainter layout(TextPainter painter) {
     return _textPainterHelper.layout(painter, this, compareChildren: true);
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    if (identical(this, other)) return true;
+    if (other.runtimeType != runtimeType) return false;
+    final OverFlowTextSpan typedOther = other;
+    return typedOther.text == text &&
+        typedOther.style == style &&
+        typedOther.recognizer == recognizer &&
+        typedOther.background == background &&
+        listEquals<TextSpan>(typedOther.children, children);
+  }
+
+  @override
+  int get hashCode =>
+      hashValues(style, text, recognizer, background, hashList(children));
+
+  @override
+  RenderComparison compareTo(TextSpan other) {
+    if (other is OverFlowTextSpan) {
+      if (other.background != background) {
+        return RenderComparison.paint;
+      }
+    }
+
+    // TODO: implement compareTo
+    return super.compareTo(other);
   }
 }